+++ libs/base/src/code/ops/mod.rs
@@ -1,0 +1,19 @@
+// devela_base::code::ops
+//
+#![doc = crate::_DOC_CODE_PANIC!()]
+//
+
+mod reexports;
+
+crate::items! { // structural access: _mods, _all
+    #[allow(unused)]
+    pub use _mods::*;
+
+    mod _mods { #![allow(unused)]
+        pub use super::reexports::*;
+    }
+    pub(super) mod _all { #![allow(unused)]
+        #[doc(inline)]
+        pub use super::_mods::*;
+    }
+}