+++ src/code/ops/reexports.rs
@@ -1,0 +1,6 @@
+// devela::code::ops::reexports
+//
+//! General reexported items, except macros and overloadable operators.
+//
+
+crate::_reexport_from!("../../../libs/base/src/code/ops/reexports.rs", _c);