+++ src/num/quant/reexports.rs
@@ -1,0 +1,11 @@
+// devela::num::quant::reexports
+//
+//!
+//
+
+#[doc(inline)] #[rustfmt::skip]
+pub use devela_base::{
+    Cycle, CycleCount,
+    Interval,
+    Sign,
+};